refactor(project structure): reorganize imports and update function signature
- Program is non-operational

Reorganized the project structure by moving `image_converter` and `image_collector` modules into a `src` directory, updating import paths accordingly. Removed `requirements.txt` as it is no longer needed in the current setup. Modified the `convert_image` function to include an `output_format` parameter, aligning with the updated `ImageConverter` class requirements. These changes improve code organization and functionality, ensuring compatibility with the new module structure and enhancing image conversion capabilities.
