--- requirements.txt
@@ -1,4 +1,0 @@
-loguru
-pytest
-Pillow
-odfpy
